Corrected get_filename() and divide_string()

+++ src/divide_string.cpp
@@ -10,8 +10,8 @@ using namespace std;
 vector<double> divide_string(string str, unsigned int number_of_values, char divider) {
 	vector<double> values;
 
-	int divider_old = -1;
-	int divider_new = -1;
+	size_t divider_old = -1;
+	size_t divider_new = -1;
 	bool end_of_string = false;
 	for(unsigned int i = 0; i < number_of_values; i++) {
 		divider_old = divider_new;
@@ -21,7 +21,7 @@ vector<double> divide_string(string str, unsigned int number_of_values, char div
 		} else {
 			values.push_back(atof(str.substr(divider_old + 1, divider_new - divider_old - 1).c_str()));
 		}
-		if((unsigned int)divider_new == str.npos) { end_of_string = true; }
+		if(divider_new == str.npos) { end_of_string = true; }
 	}
 	return values;
 }
